# packages/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enspatium API",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "enspatium_session"
      }
    },
...
  },
  "paths": {
...
    "/users/{id}": {
      "get": {
        "operationId": "getUser",
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "email",
                    "displayName",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "minLength": 1,
                    "maxLength": 320,
                    "type": "string"
                  },
                  "password": {
                    "minLength": 1,
                    "maxLength": 1024,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "email",
                    "displayName",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/namespaces/{namespaceSlug}/spaces/{spaceSlug}/audit-events": {
      "get": {
        "operationId": "listSpaceAuditEvents",
        "tags": [
          "audit"
        ],
        "parameters": [
          {
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "namespaceSlug",
            "required": true
          },
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "spaceSlug",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "actorUserId",
                      "namespaceId",
                      "spaceId",
                      "action",
                      "metadata",
                      "createdAt"
                    ],
                    "properties": {
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "actorUserId": {
                        "anyOf": [
                          {
                            "format": "uuid",
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "namespaceId": {
                        "anyOf": [
                          {
                            "format": "uuid",
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "spaceId": {
                        "anyOf": [
                          {
                            "format": "uuid",
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "action": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "space.created"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "space.updated"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "space.deleted"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "git.pushed"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "object.uploaded"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "object.deleted"
                            ]
                          }
                        ]
                      },
                      "metadata": {
                        "type": "object",
                        "additionalProperties": {}
                      },
                      "createdAt": {
                        "format": "date-time",
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "session": []
    }
  ]
}